+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Service/shelterService.java
@@ -20,7 +20,7 @@ public class shelterService {
         save(new Shelter("Protectora1","12312312","Madrid",67584637,"Protectora1@gmail.com"));//ponemos los datos tal cual el constructor
 
     }
-    public Optional<Shelter> findById(Integer id) {
+    public Optional<Shelter> findById(Long id) {
         return shelter.findById(id);
 
     }
@@ -41,7 +41,7 @@ public class shelterService {
         shelter.save(updatedShelter);
     }
 
-    public void deleteById(Integer id){//borramos usuario
+    public void deleteById(Long id){//borramos usuario
 
         this.shelter.deleteById(id);
